refactor: Organize directory
