+++ client/src/lib/collab-adapter/firebase/persistence.ts
@@ -1,7 +1,7 @@
 import * as Y from 'yjs'
 import type { Database } from 'firebase/database'
 import { ref, set, get, serverTimestamp } from 'firebase/database'
-import { buildDatabasePaths, type DatabasePathsConfig } from './config'
+import { buildDatabasePaths, type DatabasePathsConfig } from '../utils/config'
 
 // ============================================================================
 // CONSTANTS
